# openapi.yaml
openapi: "3.1.0"
info:
  title: skopos
  version: "1.0.0"
  description: >
    Skopos coordinates AI agent status, a shared blackboard, and plans across
    sessions and branches.

    REST API, dashboard, and MCP (at `/mcp`) are all served on one HTTP server
    (default :8080).

    When an API key is configured, every endpoint (reads, writes, MCP, SSE)
    requires it, sent as `Authorization: Bearer <key>`.
    When no API key is configured, authentication is disabled.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/sessions:
    get:
      summary: List sessions
      operationId: listSessions
      security: [{ BearerAuth: [] }]
      parameters:
        - name: workspace_id
          in: query
          schema: { type: string }
          description: Filter by workspace path/id.
      responses:
        "200":
          description: Session summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionSummary" }
  /api/sessions/{id}:
    get:
      summary: Get a session (with agents and events)
      operationId: getSession
      security: [{ BearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/SessionId" }
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404": { description: Session not found }
    delete:
      summary: Delete a session (cascades to its agents, events, and session-scoped blackboard entries)
      operationId: deleteSession
      security: [{ BearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/SessionId" }
      responses:
        "204": { description: Deleted }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
...
  /api/plans:
    post:
      summary: Create a plan
      operationId: createPlan
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePlanInput" }
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plan" }
        "400": { description: Invalid input, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
    get:
      summary: List plans
      operationId: listPlans
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: workspace_id, in: query, schema: { type: string } }
        - { name: branch, in: query, schema: { type: string } }
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Plan" }
  /api/plans/{id}:
    get:
      summary: Get a plan with all items and dependencies
      operationId: getPlan
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/PlanId" }]
      responses:
        "200":
          description: Plan detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plan" }
        "404": { description: Plan not found }
    patch:
      summary: Update a plan (name, description, status)
      operationId: updatePlan
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/PlanId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePlanInput" }
      responses:
        "204": { description: Updated }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
        "404": { description: Plan not found }
    delete:
      summary: Delete a plan (cascades to items and dependencies)
      operationId: deletePlan
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/PlanId" }]
      responses:
        "204": { description: Deleted }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
        "404": { description: Plan not found }
...
  /api/plans/{id}/dependencies:
    post:
      summary: Add a plan-to-plan dependency (auto-blocks the dependent plan)
      operationId: addPlanDependency
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/PlanId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [depends_on_plan_id]
              properties:
                depends_on_plan_id: { type: string }
      responses:
        "200":
          description: Updated dependent plan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plan" }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
        "404": { description: Not found }
        "409": { description: Dependency cycle detected }
  /api/plans/{id}/dependencies/{depends_on_id}:
    delete:
      summary: Remove a plan-to-plan dependency
      operationId: removePlanDependency
      security: [{ BearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/PlanId" }
        - { name: depends_on_id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "401": { description: Unauthorized }
        "403": { description: "Workspace out of scope for this credential (or root key required)" }
        "404": { description: Not found }
components:
  parameters:
    SessionId:  { name: id, in: path, required: true, schema: { type: string } }
...
    PlanId:     { name: id, in: path, required: true, schema: { type: string } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        Send an API key as the bearer token: the root key (server config, full
        access and exclusive key/workspace management) or a scoped key minted
        via POST /api/keys. Scoped keys are limited to their workspace list —
        explicit-workspace operations outside the scope return 403 (listing
        the accessible workspaces), by-id operations return 404, and writes
        require workspace_id for every principal including root.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Status:
      type: string
      description: >
        Agent status. `stuck` and `orphaned` are set exclusively by the health
        checker; agents must not self-report them.
      enum:
        - pending
        - thinking
        - planning
        - running
        - editing
        - testing
        - waiting
        - blocked
        - paused
        - handoff
        - succeeded
        - failed
        - cancelled
        - stuck
        - orphaned
...
    AgentState:
      type: object
      properties:
        session_id: { type: string }
        agent_id: { type: string }
        agent_type: { type: string }
        workspace: { type: string }
        status: { $ref: "#/components/schemas/Status" }
        progress: { type: integer }
        step_current: { type: integer }
        step_total: { type: integer }
        message: { type: string }
        snippet: { type: string }
        metadata: { type: object, additionalProperties: true }
        updated_at: { type: string, format: date-time }
        original_status: { $ref: "#/components/schemas/Status" }
        stuck_at: { type: string, format: date-time }
        git_branch: { type: string }
    Event:
      allOf:
        - $ref: "#/components/schemas/AgentState"
        - type: object
          properties:
            id: { type: string }
            created_at: { type: string, format: date-time }
    SessionSummary:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        workspace: { type: string }
        status: { $ref: "#/components/schemas/Status" }
        agent_count: { type: integer }
        started_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SessionDetail:
      allOf:
        - $ref: "#/components/schemas/SessionSummary"
        - type: object
          properties:
            agents:
              type: array
              items: { $ref: "#/components/schemas/AgentState" }
            events:
              type: array
              items: { $ref: "#/components/schemas/Event" }
...
    PlanStatus:
      type: string
      enum: [active, completed, archived, blocked]
    ItemStatus:
      type: string
      enum: [pending, in_progress, done, blocked]
    PlanItem:
      type: object
      properties:
        id: { type: string }
        plan_id: { type: string }
        title: { type: string }
        description: { type: string }
        phase: { type: string }
        status: { $ref: "#/components/schemas/ItemStatus" }
        position: { type: integer }
        claimed_by_agent_id: { type: string }
        depends_on:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Plan:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        branch_name: { type: string }
        workspace_id: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/PlanStatus" }
        author_agent_id: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/PlanItem" }
        depends_on:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreatePlanInput:
      type: object
      required: [name, author_agent_id]
      properties:
        name: { type: string }
        branch_name: { type: string }
        workspace_id: { type: string }
        description: { type: string }
        author_agent_id: { type: string }
    UpdatePlanInput:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/PlanStatus" }
...